pyproject.toml
[tool.poetry]
name = "dataenwatnu"
version = "0.1.0"
description = ""
authors = ["Eddy van den Aker <eddy.v.d.aker@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
python-dotenv = "^0.19.2"
psycopg2 = "^2.9.2"
Flask-WTF = "^1.0.0"
email-validator = "^1.1.3"
pandas = "^1.3.4"
matplotlib = "^3.5.0"
Flask-Mail = "^0.9.1"
Flask-Bootstrap = "^3.3.7"
Flask-Login = "^0.5.0"
Flask-HTTPAuth = "^4.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.4.0"
autopep8 = "^1.6.0"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
